split and re-group game module files
